+++ cfunits/test/test_Units.py
@@ -2,34 +2,39 @@ import math
 import os
 import unittest
 
+import numpy
+
 import cfunits
 from cfunits import Units
 
 
 class UnitsTest(unittest.TestCase):
     def test_Units___eq__(self):
-        self.assertTrue(Units('')==Units(''))
-        self.assertTrue(Units('18')==Units('18'))
-        self.assertTrue(Units('1')==Units('1'))
-        self.assertTrue(Units('m')==Units('m'))
-        self.assertTrue(Units('m')==Units('metres'))
-        self.assertTrue(Units('m')==Units('meTRES'))
+        self.assertEqual(Units(''), Units(''))
+        self.assertEqual(Units('18'), Units('18'))
+        self.assertEqual(Units('1'), Units('1'))
+        self.assertEqual(Units('m'), Units('m'))
+        self.assertEqual(Units('m'), Units('metres'))
+        self.assertEqual(Units('m'), Units('meTRES'))
 
-        self.assertTrue(
-            Units('days since 2000-1-1')==Units('d since 2000-1-1 0:0'))
-        self.assertTrue(
-            Units('days since 2000-1-1')!=Units('h since 1234-1-1 0:0'))
+        self.assertEqual(Units('days since 2000-1-1'),
+                         Units('d since 2000-1-1 0:0'))
+        self.assertNotEqual(Units('days since 2000-1-1'),
+                            Units('h since 1234-1-1 0:0'))
         
-        self.assertTrue(Units('days since 2000-1-1')==Units('d since 2000-1-1 0:0', calendar='gregorian'))
-        self.assertTrue(Units('days since 2000-1-1')==Units('d since 2000-1-1 0:0', calendar='standard'))
+        self.assertEqual(Units('days since 2000-1-1'),
+                         Units('d since 2000-1-1 0:0', calendar='gregorian'))
+        self.assertEqual(Units('days since 2000-1-1'),
+                         Units('d since 2000-1-1 0:0', calendar='standard'))
         
-        self.assertTrue(Units(calendar='noleap')==Units(calendar='noleap'))
-        self.assertTrue(Units(calendar='noleap')==Units(calendar='365_day'))
-        self.assertTrue(Units(calendar='nOLEAP')==Units(calendar='365_dAY'))
+        self.assertEqual(Units(calendar='noleap'), Units(calendar='noleap'))
+        self.assertEqual(Units(calendar='noleap'), Units(calendar='365_day'))
+        self.assertEqual(Units(calendar='nOLEAP'), Units(calendar='365_dAY'))
         
-        self.assertTrue(Units('days since 2000-1-1', calendar='all_leap')==Units('d since 2000-1-1 0:0', calendar='366_day'))
-        self.assertTrue(Units('days since 2000-1-1', calendar='all_leap')!=Units('h since 2000-1-1 0:0', calendar='366_day'))    
-
+        self.assertEqual(Units('days since 2000-1-1', calendar='all_leap'),
+                         Units('d since 2000-1-1 0:0', calendar='366_day'))
+        self.assertNotEqual(Units('days since 2000-1-1', calendar='all_leap'),
+                            Units('h since 2000-1-1 0:0', calendar='366_day'))
         
     def test_Units_equivalent(self):
         self.assertTrue(Units('').equivalent(Units('')))
@@ -61,92 +66,146 @@ class UnitsTest(unittest.TestCase):
 
         u = Units('days since 2000-02-02', calendar='standard')
         v = Units('months since 2000-02-02', calendar='standard')
-        self.assertFalse(u == v)
+        self.assertNotEqual(u, v)
 
         u = Units('days since 2000-02-02', calendar='standard')
         v = Units('months since 2000-02-02', calendar='gregorian')
-        self.assertFalse(u == v)
+        self.assertNotEqual(u, v)
 
     def test_Units_conform(self):
-        self.assertTrue(Units.conform(0.5, Units('km'), Units('m')) == 500)
-        self.assertTrue(
-            Units.conform(360, Units('second'), Units('minute')) == 6)
+        self.assertEqual(Units.conform(0.5, Units('km'), Units('m')), 500)
+        
+        self.assertEqual(
+            Units.conform(360, Units('second'), Units('minute')), 6)
+
+        x = Units.conform([360], Units('second'), Units('minute'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, 6))
+        
+        x = Units.conform((360, 720), Units('second'), Units('minute'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, [6, 12]))
+        
+        x = Units.conform([360.0, 720.0], Units('second'), Units('minute'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, [6, 12]))
+        
+        x = Units.conform([[360, 720]], Units('second'), Units('minute'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, [[6, 12]]))
+
+        v = numpy.array([360.0, 720.0])
+        x = Units.conform(v, Units('second'), Units('minute'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, [6, 12]), x)
+        
+        v = numpy.array([360, 720])
+        x = Units.conform(v, Units('second'), Units('minute'), inplace=True)
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, [6, 12]))
+        self.assertTrue(numpy.allclose(x, v))    
+        
+        x = Units.conform(35, Units('degrees_C'), Units('degrees_F'))
+        self.assertIsInstance(x, float)
+        self.assertTrue(numpy.allclose(x, 95))
+        
+        x = Units.conform([35], Units('degrees_C'), Units('degrees_F'))
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, 95))
+        
+        x = Units.conform(35, Units('degrees_C'), Units('degrees_F'),
+                          inplace=True)
+        self.assertIsInstance(x, float)
+        self.assertTrue(numpy.allclose(x, 95))
+        
+        x = Units.conform([35], Units('degrees_C'), Units('degrees_F'),
+                          inplace=True)
+        self.assertIsInstance(x, numpy.ndarray)
+        self.assertEqual(x.dtype, numpy.dtype('float64'))
+        self.assertTrue(numpy.allclose(x, 95))
+        
         with self.assertRaises(ValueError):
             Units.conform(1, Units('m'), Units('second'))
 
     def test_Units_BINARY_AND_UNARY_OPERATORS(self):
-        self.assertTrue((Units('m')*2)    ==Units('2m'))
-        self.assertTrue((Units('m')/2)    ==Units('0.5m'))
-        self.assertTrue((Units('m')//2)   ==Units('0.5m'))
-        self.assertTrue((Units('m')+2)    ==Units('m @ -2'))
-        self.assertTrue((Units('m')-2)    ==Units('m @ 2'))
-        self.assertTrue((Units('m')**2)   ==Units('m2'))
-        self.assertTrue((Units('m')**-2)  ==Units('m-2'))
-        self.assertTrue((Units('m2')**0.5)==Units('m'))
+        self.assertEqual(Units('m') * 2, Units('2m'))
+        self.assertEqual(Units('m') / 2, Units('0.5m'))
+        self.assertEqual(Units('m') // 2, Units('0.5m'))
+        self.assertEqual(Units('m') + 2, Units('m @ -2'))
+        self.assertEqual(Units('m') - 2, Units('m @ 2'))
+        self.assertEqual(Units('m') ** 2, Units('m2'))
+        self.assertEqual(Units('m') ** -2, Units('m-2'))
+        self.assertEqual(Units('m2') ** 0.5, Units('m'))
     
         u = Units('m')
         v = u
         u *= 2
-        self.assertTrue(u==Units('2m'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('2m'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u /= 2
-        self.assertTrue(u==Units('0.5m'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('0.5m'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u //= 2
-        self.assertTrue(u==Units('0.5m'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('0.5m'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u += 2
-        self.assertTrue(u==Units('m @ -2'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('m @ -2'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u -= 2
-        self.assertTrue(u==Units('m @ 2'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('m @ 2'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u **= 2
-        self.assertTrue(u==Units('m2'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('m2'))
+        self.assertNotEqual(u, v)
     
-        self.assertTrue((2*Units('m')) ==Units('2m'))
-        self.assertTrue((2/Units('m')) ==Units('2 m-1'))
-        self.assertTrue((2//Units('m'))==Units('2 m-1'))
-        self.assertTrue((2+Units('m')) ==Units('m @ -2'))
-        self.assertTrue((2-Units('m')) ==Units('-1 m @ -2'))
+        self.assertEqual(2 * Units('m'), Units('2m'))
+        self.assertEqual(2 / Units('m'), Units('2 m-1'))
+        self.assertEqual(2 // Units('m'),Units('2 m-1'))
+        self.assertEqual(2 + Units('m'), Units('m @ -2'))
+        self.assertEqual(2 - Units('m'), Units('-1 m @ -2'))
     
-        self.assertTrue((Units('m')*Units('2m')) ==Units('2 m2'))
-        self.assertTrue((Units('m')/Units('2m')) ==Units('0.5'))
-        self.assertTrue((Units('m')//Units('2m'))==Units('0.5'))
+        self.assertEqual(Units('m') * Units('2m'), Units('2 m2'))
+        self.assertEqual(Units('m') / Units('2m'), Units('0.5'))
+        self.assertEqual(Units('m') // Units('2m'), Units('0.5'))
     
         u = Units('m')
         v = u
         u *= u
-        self.assertTrue(u==Units('m2'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('m2'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u /= u
-        self.assertTrue(u==Units('1'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('1'))
+        self.assertNotEqual(u, v)
         u = Units('m')
         v = u
         u //= u
-        self.assertTrue(u==Units('1'))
-        self.assertTrue(u!=v)
+        self.assertEqual(u, Units('1'))
+        self.assertNotEqual(u, v)
     
-        self.assertTrue(Units('m').log(10)    == Units('lg(re 1 m)'))
-        self.assertTrue(Units('m').log(2)     == Units('lb(re 1 m)'))
-        self.assertTrue(Units('m').log(math.e)== Units('ln(re 1 m)'))
-        self.assertTrue(
-            Units('m').log(1.5)   == Units('2.46630346237643 ln(re 1 m)'))    
-
+        self.assertEqual(Units('m').log(10), Units('lg(re 1 m)'))
+        self.assertEqual(Units('m').log(2), Units('lb(re 1 m)'))
+        self.assertEqual(Units('m').log(math.e), Units('ln(re 1 m)'))
+        self.assertEqual(Units('m').log(1.5),
+                         Units('2.46630346237643 ln(re 1 m)'))    
 
     def test_Units_isvalid(self):
         self.assertTrue(Units('m').isvalid)